Deletes HoverItemsContainer component and localizes it within NavMenu
